// ErpDashboard/src/app/models/gibInbox.ts
export class GibInbox {
    uuid: string;
    invoiceId: string;
    profile: string;
    typeCode: string;
    issueDate: string;
    envelopeId: string;
    envelopeDate: string;
    date: string;
    partyId: string;
    partyName: string;
    taxSchemeName: string;
    payableAmount: string;
    currency: string;
    statusCode: string;
    statusDescription: string;
    applicationResponseId: string;
    applicationResponse: string;
    applicationResponseDate: string;
    isCancelled: string;
    cancelDate: string;
    cancelReason: string;

}